[Feat -> Dev] 스텁 데이터 작성 (#23)
* #12 - 글로벌 ResponseDto 작성

- MultiResponseDto 작성
- PageInfo 작성
- SingleResponseDto 작성

* #5 - push 알림 테스트

* #5 - 게시글 도메인 DTO 작성

- ArticleDto 작성
- ArticleCommentDto 작성

* #19 - 스텁 데이터 만들기

- 회원 스텁 데이터 작성
- 공연 스텁 데이터 작성
- 게시글 스텁 데이터 작성

* #19 - 스텁 데이터 만들기

- 회원 스텁 데이터 작성
- 공연 스텁 데이터 작성
- 게시글 스텁 데이터 작성

* #19 - 스텁 데이터 만들기

- 회원 스텁 데이터 작성
- 공연 스텁 데이터 작성
- 게시글 스텁 데이터 작성

+++ server/src/main/java/codestates/frogroup/indiego/domain/show/dto/ShowCommentDto.java
@@ -1,12 +1,8 @@
 package codestates.frogroup.indiego.domain.show.dto;
 
-import codestates.frogroup.indiego.domain.member.entity.Member;
-import codestates.frogroup.indiego.domain.show.entity.Show;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-
-import javax.persistence.*;
 
 public class ShowCommentDto {
 
@@ -15,7 +11,7 @@ public class ShowCommentDto {
     public static class Post{
         private Long id;
         private ShowDto.Post show;
-        private Integer score;
+        private Double score;
         private String comment;
     }
 
@@ -25,7 +21,7 @@ public class ShowCommentDto {
     public static class Patch{
         private Long id;
         private ShowDto.Patch show;
-        private Integer score;
+        private Double score;
         private String comment;
     }
 
@@ -35,7 +31,7 @@ public class ShowCommentDto {
     public static class Response{
         private Long id;
         private ShowDto.Response show;
-        private Integer score;
+        private Double score;
         private String comment;
     }
 }
